authentication

+++ database/migrations/2023_06_08_211519_create_posts_table.php
@@ -15,13 +15,12 @@ return new class extends Migration
             $table->id();
             $table->unsignedBigInteger('user_id');
             $table->foreign('user_id')->references('id')->on('users');
-            $table->unsignedBigInteger('professional_id');
+            $table->integer('professional_id');
             $table->foreign('professional_id')->references('id')->on('professionals');
             $table->text('content');
             $table->string('image')->nullable();
             $table->timestamps();
         });
-
     }
 
     /**
